requirements.txt
azure-functions
cryptography
pyjks